fix update

+++ app/Http/Controllers/ForemanWorkerController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ForemanWorkers\StoreWorkerRequest;
 use App\Http\Requests\ForemanWorkers\UpdateWorkerRequest;
 use App\Models\Worker;
+use App\Models\User;
 use App\Services\ForemanWorkerService;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
@@ -23,6 +24,7 @@ class ForemanWorkerController extends Controller
 
     public function store(StoreWorkerRequest $request)
     {
+        abort_unless(in_array($request->user()->role, [User::ROLE_HR, User::ROLE_HEAD_ADMIN], true), 403);
         $this->foremanWorkerService->createWorker($request->user(), $request->validated());
 
         return redirect()
@@ -32,6 +34,7 @@ class ForemanWorkerController extends Controller
 
     public function update(UpdateWorkerRequest $request, Worker $worker)
     {
+        abort_unless(in_array($request->user()->role, [User::ROLE_HR, User::ROLE_HEAD_ADMIN], true), 403);
         $this->foremanWorkerService->updateWorker($request->user(), $worker, $request->validated());
 
         return redirect()
@@ -41,6 +44,7 @@ class ForemanWorkerController extends Controller
 
     public function destroy(Request $request, Worker $worker)
     {
+        abort_unless(in_array($request->user()->role, [User::ROLE_HR, User::ROLE_HEAD_ADMIN], true), 403);
         $this->foremanWorkerService->deleteWorker($request->user(), $worker);
 
         return redirect()
